refactor(frontend): rename pages folders
