+++ frontend/src/App.tsx
@@ -4,8 +4,10 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { createTheme, ThemeProvider } from "@mui/material";
 import StyledEngineProvider from "@mui/material/StyledEngineProvider";
 
-import { CityPage } from "src/pages/citypage";
-import { SearchPage } from "src/pages/searchpage";
+import { CityPage } from "src/pages/city";
+import { SearchPage } from "src/pages/search";
+
+import { useTranslation, Trans } from 'react-i18next';
 
 import "./index.css";
 
@@ -13,6 +15,7 @@ import "./i18n";
 
 function App() {
   const theme = createTheme({});
+  const { t } = useTranslation();
 
   return (
     <StyledEngineProvider injectFirst>
