+++ frontend/src/pages/city/styles.css
@@ -1,14 +1,11 @@
-@import 'src/components/styles/variables/color.scss';
-
 .city-page {
   display: flex;
   width: 100%;
   height: 100dvh;
   justify-content: space-around;
   flex-direction: column;
-  background-image: url('./images/background.svg');
+  background-image: url("./images/background.svg");
   background-repeat: no-repeat;
-  background-color: $bg-color;
   background-size: 100%;
 }
 
@@ -30,7 +27,6 @@
 .city-page-main-card {
   display: flex;
   width: 45%;
-  background-color: $bg-color;
   border-radius: 50px;
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
   flex-direction: column;
@@ -68,7 +64,7 @@
   width: 80%;
 }
 
-.city-page-main-card-image {
+.city-page-main-cadr-image {
   display: flex;
   width: 100%;
 }
@@ -90,3 +86,5 @@
   display: flex;
   width: 50%;
 }
+
+/*# sourceMappingURL=styles.css.map */
