+++ frontend/src/pages/index.ts
@@ -1,1 +1,1 @@
-export * from './searchpage';
+export * from './search';
