+++ frontend/src/pages/search/styles.css
